+++ packages/strum-react/src/components/Radio/Radio.tsx
@@ -1,12 +1,16 @@
 import * as RadixLabel from '@radix-ui/react-label';
 import * as RadixRadio from '@radix-ui/react-radio-group';
 import * as React from 'react';
+import { Box } from '../Box';
 import { Stack } from '../Stack';
+import { Text } from '../Text';
 import * as styles from './Radio.css';
 
 type RadioGroupProps = {
+  /** String to display if an error is present */
+  error?: string;
   /** Description of the field */
-  label: string;
+  label?: string;
   /** onChange callback for a controlled input */
   onChange?: RadixRadio.RadioGroupProps['onValueChange'];
   /** The value of a controlled input group */
@@ -14,41 +18,68 @@ type RadioGroupProps = {
 } & RadixRadio.RadioGroupProps;
 
 export const RadioGroup = React.forwardRef<HTMLDivElement, RadioGroupProps>(
-  ({ children, label, ...primitiveProps }, ref) => {
+  ({ children, error, label, ...primitiveProps }, ref) => {
+    const errorId = error ? `${label}-error` : undefined;
+
     return (
-      <RadixRadio.Root aria-label={label} ref={ref} {...primitiveProps}>
-        <Stack alignItems="flex-start" direction="vertical">
-          {children}
-        </Stack>
-      </RadixRadio.Root>
+      <>
+        <RadixRadio.Root
+          className={error ? styles.radioGroupWithErrorStyle : undefined}
+          ref={ref}
+          {...primitiveProps}
+        >
+          {label && (
+            <Box marginBottom="4">
+              <Text color="neutral12" size="small">
+                {label}
+              </Text>
+            </Box>
+          )}
+          <Stack alignItems="flex-start" direction="vertical">
+            {children}
+
+            {/* Feedback text */}
+            {error && (
+              <Box id={errorId}>
+                <Text color="error11" size="small">
+                  {error}
+                </Text>
+              </Box>
+            )}
+          </Stack>
+        </RadixRadio.Root>
+      </>
     );
   },
 );
 
-RadioGroup.displayName = 'RadioGroup';
+RadioGroup.displayName = 'Radio';
 
 type RadioItemProps = {
-  /** Ties the label to the input field */
-  id: string;
+  /** Ties the label to the input field. The `value` field is used if not present. */
+  id?: string;
   /** Text of the option */
   label: string;
 } & RadixRadio.RadioGroupItemProps;
 
 export const RadioItem = React.forwardRef<HTMLButtonElement, RadioItemProps>(
-  ({ id, label, value, ...primitiveProps }, ref) => {
+  ({ disabled, id, label, value, ...primitiveProps }, ref) => {
+    const labelId = id || value;
+
     return (
       <Stack alignItems="center" direction="horizontal">
         <RadixRadio.Item
           className={styles.radioStyle}
-          id={id}
+          disabled={disabled}
+          id={labelId}
           ref={ref}
           value={value}
           {...primitiveProps}
         >
           <RadixRadio.Indicator className={styles.indicatorStyle} />
         </RadixRadio.Item>
-        <RadixLabel.Label className={styles.radioLabelStyle} htmlFor={id}>
-          {label}
+        <RadixLabel.Label className={styles.radioLabelStyle} htmlFor={labelId}>
+          <Text color={disabled ? 'neutral11' : 'neutral12'}>{label}</Text>
         </RadixLabel.Label>
       </Stack>
     );