+++ packages/strum-react/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `<Badge />` component
 - `<Checkbox />` component
+- `<RadioGroup />` and `<RadioItem />` components
 - New `error` prop on the `<Select />` component
 - New `borderRadius` prop on the `<Button />` component